+++ src/Controllers/ProductController.php
@@ -17,6 +17,7 @@ use Jtl\Connector\Core\Model\ProductI18n as ProductI18nModel;
 use Jtl\Connector\Core\Model\QueryFilter;
 use Jtl\Connector\Core\Model\TaxRate;
 use JtlWooCommerceConnector\Controllers\Product\Product2CategoryController;
+use JtlWooCommerceConnector\Controllers\Product\ProductAdvancedCustomFieldsController;
 use JtlWooCommerceConnector\Controllers\Product\ProductB2BMarketFieldsController;
 use JtlWooCommerceConnector\Controllers\Product\ProductDeliveryTimeController;
 use JtlWooCommerceConnector\Controllers\Product\ProductGermanizedFieldsController;
@@ -221,6 +222,10 @@ class ProductController extends AbstractBaseController implements
                 }
             }
 
+            if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_ADVANCED_CUSTOM_FIELDS)) {
+                (new ProductAdvancedCustomFieldsController($this->db, $this->util))->pullData($productModel, $product);
+            }
+
             $products[] = $productModel;
         }
 
@@ -360,6 +365,10 @@ class ProductController extends AbstractBaseController implements
             || SupportedPlugins::isActive(SupportedPlugins::PLUGIN_RANK_MATH_SEO)
         ) {
             (new ProductMetaSeoController($this->db, $this->util))->pushData((int)$newPostId, $tmpI18n);
+        }
+
+        if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_ADVANCED_CUSTOM_FIELDS)) {
+            (new ProductAdvancedCustomFieldsController($this->db, $this->util))->pushData($model);
         }
 
         \remove_filter('content_save_pre', 'wp_filter_post_kses');